Restructure files in the iOS target
